Add test for the activities declarations
adds the test only for the public ativities.
the tests are commented (xscenario) because for some reason
when the page clicks on a tab (Activities declarations in this case)
the form on the hidden tab becomes empty, so the validator prevent
the submit button from sending the form.

Im still working on it