Use "&&" instead of "and"
The operator "and" in ruby is used for flow control; it isn't a logical
operator and its usage might lead to unexpected results.

For example, executing `a = true and false` returns false (as expected)
but after executing it `a` is assigned to true, which isn't what most
people would expect.